extract useAutoConnect hook
