+++ packages/thirdweb/src/exports/react-native.ts
@@ -79,3 +79,10 @@ export const useSendTransaction = useSendTransactionCore;
 export { useSendBatchTransaction } from "../react/core/hooks/contract/useSendBatchTransaction.js";
 export { useSendAndConfirmTransaction } from "../react/core/hooks/contract/useSendAndConfirmTransaction.js";
 export { useEstimateGasCost } from "../react/core/hooks/contract/useEstimateGasCost.js";
+
+// Components
+export {
+  AutoConnect,
+  type AutoConnectProps,
+} from "../react/core/hooks/connection/AutoConnect.js";
+export { useAutoConnect } from "../react/core/hooks/connection/useAutoConnect.js";
