+++ packages/thirdweb/src/react/core/hooks/connection/AutoConnect.tsx
@@ -2,22 +2,10 @@
 import { useEffect } from "react";
 import type { ThirdwebClient } from "../../../../client/client.js";
 import type { Wallet } from "../../../../wallets/interfaces/wallet.js";
-import {
-  getLastConnectedChain,
-  getStoredActiveWalletId,
-  getStoredConnectedWalletIds,
-} from "../../../../wallets/manager/index.js";
 import type { SmartWalletOptions } from "../../../../wallets/smart/types.js";
 import type { AppMetadata } from "../../../../wallets/types.js";
-import { connectionManager } from "../../connectionManager.js";
-import { getStorage } from "../../storage.js";
-import { timeoutPromise } from "../../utils/timeoutPromise.js";
-import {
-  useConnect,
-  useSetActiveWalletConnectionStatus,
-} from "../wallets/wallet-hooks.js";
-
-let autoConnectAttempted = false;
+import { useSetActiveWalletConnectionStatus } from "../wallets/wallet-hooks.js";
+import { useAutoConnect } from "./useAutoConnect.js";
 
 export type AutoConnectProps = {
   /**
@@ -157,103 +145,7 @@ export type AutoConnectProps = {
  * @component
  */
 export function AutoConnect(props: AutoConnectProps) {
-  const setConnectionStatus = useSetActiveWalletConnectionStatus();
-  const { connect } = useConnect({
-    client: props.client,
-    accountAbstraction: props.accountAbstraction,
-  });
-  const { isAutoConnecting } = connectionManager;
-  const { wallets, onConnect } = props;
-  const timeout = props.timeout ?? 15000;
-  // get the supported wallets from thirdweb provider
-  // check the storage for last connected wallets and connect them all
-  // check the storage for last active wallet and set it as active
-  useEffect(() => {
-    if (autoConnectAttempted) {
-      return;
-    }
-
-    const asyncLocalStorage = getStorage();
-    autoConnectAttempted = true;
-
-    const startAutoConnect = async () => {
-      const [lastConnectedWalletIds, lastActiveWalletId] = await Promise.all([
-        getStoredConnectedWalletIds(asyncLocalStorage),
-        getStoredActiveWalletId(asyncLocalStorage),
-      ]);
-
-      // if no wallets were last connected
-      if (!lastConnectedWalletIds) {
-        return;
-      }
-
-      const lastConnectedChain = await getLastConnectedChain(asyncLocalStorage);
-
-      async function handleWalletConnection(wallet: Wallet) {
-        return wallet.autoConnect({
-          client: props.client,
-          chain: lastConnectedChain ?? undefined,
-        });
-      }
-
-      const activeWallet =
-        lastActiveWalletId && wallets.find((w) => w.id === lastActiveWalletId);
-
-      if (activeWallet) {
-        try {
-          setConnectionStatus("connecting"); // only set connecting status if we are connecting the last active EOA
-          await timeoutPromise(handleWalletConnection(activeWallet), {
-            ms: timeout,
-            message: `AutoConnect timeout : ${timeout}ms limit exceeded.`,
-          });
-
-          // connected wallet could be activeWallet or smart wallet
-          const connectedWallet = await connect(activeWallet);
-
-          if (connectedWallet) {
-            if (onConnect) {
-              try {
-                onConnect(connectedWallet);
-              } catch {
-                // ignore
-              }
-            }
-          } else {
-            setConnectionStatus("disconnected");
-          }
-        } catch (e) {
-          console.error("Failed to auto connect last active wallet");
-          console.error(e);
-          setConnectionStatus("disconnected");
-        }
-      } else {
-        setConnectionStatus("disconnected");
-      }
-
-      // then connect wallets that were last connected but were not set as active
-      const otherWallets = wallets.filter(
-        (w) =>
-          w.id !== lastActiveWalletId && lastConnectedWalletIds.includes(w.id),
-      );
-
-      for (const wallet of otherWallets) {
-        try {
-          await handleWalletConnection(wallet);
-          connectionManager.addConnectedWallet(wallet);
-        } catch (e) {
-          console.error("Failed to auto connect a non-active connected wallet");
-          console.error(e);
-        }
-      }
-    };
-
-    (async () => {
-      isAutoConnecting.setValue(true);
-      await startAutoConnect();
-      isAutoConnecting.setValue(false);
-    })();
-  });
-
+  useAutoConnect(props);
   return <></>;
 }
